Increase gunicorn worker timeout
Among other things the gunicorn worker manages restarting systemd
services. Most of these services have TimeoutStopSec set to 30.
This means that if any of the services are not responding then
systemd will wait 30s for before killing the service. However the
gunicorn worker also has a timeout of 30s meaning that the gunicorn
worker ends up getting killed as well. This change increases the
worker timeout from 30s to 2 minutes.

This situation seems to occur most often during a microstack
reset. The model hosting rabbitmq is removed and the nova-compute
seems to not respond to shutdown requests while reporting
EHOSTUNREACH errors.