Stop setting 'unset' snap config values
The snap uses pydantic to validate data. Setting a key to have a
value of UNSET ('') in some cases fails this validation. For
example the rabbitmq.url value must conform to pydantic AnyURL which
does not permit and empty string. To avoid this stop setting
snap config values to UNSET.