Use fallback for determining interface for local ip
In some cases, the netifaces does not properly detect the default
gateway for the node and returns an empty dict. However, there is
a default gateway actually set. This adds a fallback path to parse
the contents of the /proc/net/route table and determine which
interface to use based on which flags are set on the routes.

Closes-Bug: #2030497

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>
(cherry picked from commit 51b2584268320cccc626bf6366f71c1c9f82a622)